+++ src/rest.ts
@@ -1,11 +1,10 @@
 export {ResponseType, LocationHeaderAndJsonResponse, LocationHeaderResponse, BlobAndFilenameResponse} from './rest/responseType';
-export {Cache} from './rest/cache';
 export {DELETE, GET, HEAD, POST, PUT, RESTClient} from './rest/common';
 export {RestTransferStateService} from './transferState/restTransferState.service';
 export {AdditionalInfoPropertyDescriptor, ParametersMetadata, RestCaching, RestDisabledInterceptors, RestFullHttpResponse, RestHttpHeaders, RestMethod, RestParameters, RestReportProgress, RestResponseTransform, RestResponseType, KeyIndex, ParametersTransformMetadata, RestMiddleware, RestMiddlewareRunMethod, ɵRESTClient, RestMethodMiddlewares, NotType, BuildMiddlewaresFn, ParametersMiddlewaresMetadata, ParametersTransformsObj} from './rest/rest.interface';
 export {HTTP_HEADER_ACCEPT, HTTP_HEADER_CONTENT_TYPE} from './rest/constants';
-export {Body, BaseUrl, DefaultHeaders, Header, Headers, JsonContentType, ParameterTransform, Path, Produces, Query, QueryObject, ResponseTransform, FullHttpResponse, ReportProgress, DisableInterceptor, ProgressIndicatorGroup, AcceptAny} from './rest/decorators';
+export {Body, BaseUrl, DefaultHeaders, Header, Headers, JsonContentType, ParameterTransform, Path, Produces, Query, QueryObject, ResponseTransform, FullHttpResponse, ReportProgress, DisableInterceptor, ProgressIndicatorGroup, AcceptAny, Cache} from './rest/decorators';
 export {REST_METHOD_MIDDLEWARES, REST_MIDDLEWARES_ORDER} from './rest/tokens';
 export {BASIC_DEFAULT_REST_METHOD_MIDDLEWARES, BASIC_DEFAULT_REST_MIDDLEWARES_ORDER} from './rest/defaults';
-export {ReportProgressMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, ResponseTypeMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, HeaderParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeadersMiddleware} from './rest/middlewares';
+export {ReportProgressMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, ResponseTypeMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, HeaderParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeadersMiddleware, CacheMiddleware} from './rest/middlewares';
 export {buildMiddlewares, getType, isNotType, not} from './rest/utils';