+++ src/rest/decorators/index.ts
@@ -11,3 +11,4 @@ export * from './progressIndicatorGroup.decorator';
 export * from './reportProgress.decorator';
 export * from './responseTransform.decorator';
 export * from './acceptAny.decorator';
+export * from './cache.decorator';